@@ -1,7 +1,7 @@
 {
   "name": "fluig-monitor",
   "description": "Application for monitoring Fluig environments.",
-  "version": "0.3.0",
+  "version": "0.4.0",
   "scripts": {
     "build": "concurrently \"npm run build:main\" \"npm run build:renderer\"",
     "build:main": "cross-env NODE_ENV=production TS_NODE_TRANSPILE_ONLY=true webpack --config ./.erb/configs/webpack.config.main.prod.ts && ts-node ./scripts/copySchema.ts",
@@ -164,7 +164,7 @@
     "@types/node-forge": "^1.3.1",
     "@types/node-schedule": "^2.1.0",
     "@types/react": "^17.0.33",
-    "@types/react-dom": "^17.0.10",
+    "@types/react-dom": "^18.0.10",
     "@types/react-router": "^5.1.17",
     "@types/react-router-dom": "^5.3.2",
     "@types/react-test-renderer": "^17.0.1",
@@ -180,7 +180,7 @@
     "css-loader": "^6.5.0",
     "css-minimizer-webpack-plugin": "^3.1.1",
     "detect-port": "^1.3.0",
-    "electron": "^18.0.4",
+    "electron": "^22.0.2",
     "electron-builder": "23.0.8",
     "electron-devtools-installer": "^3.2.0",
     "electron-notarize": "^1.1.1",
@@ -233,6 +233,7 @@
     "@popperjs/core": "^2.11.0",
     "@prisma/client": "^4.6.1",
     "axios": "^0.24.0",
+    "compressing": "^1.7.0",
     "electron-debug": "^3.2.0",
     "electron-log": "^4.4.8",
     "electron-store": "^8.1.0",
@@ -244,11 +245,11 @@
     "node-forge": "^1.3.1",
     "node-schedule": "^2.1.0",
     "oauth-1.0a": "^2.2.6",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-i18next": "^11.16.7",
     "react-icons": "^4.3.1",
-    "react-router-dom": "^5.3.0",
+    "react-router-dom": "^6.6.2",
     "recharts": "^2.1.16",
     "regenerator-runtime": "^0.13.9",
     "request": "^2.88.2"
