@@ -2,6 +2,7 @@
 import path from 'path';
 import * as fs from 'fs';
 import log from 'electron-log';
+import compressing from 'compressing';
 import { isDevelopment } from './globalConstants';
 import getAppDataFolder from './fsUtils';
 
@@ -9,13 +10,13 @@ import getAppDataFolder from './fsUtils';
  * Archives the app log file on a daily basis with a custom name, preventing it from being too large.
  * @since 0.1.0
  */
-export default function rotateLogFile(): void {
+export default async function rotateLogFile(): Promise<void> {
   try {
     const today = new Date();
     const filePath = path.resolve(
       getAppDataFolder(),
       'logs',
-      isDevelopment ? 'app.dev.log' : 'app.log'
+      isDevelopment ? 'fluig-monitor.dev.log' : 'fluig-monitor.log'
     );
     const yesterday = new Date().setDate(today.getDate() - 1);
     const yesterdayFileFormat = new Date(yesterday)
@@ -35,18 +36,27 @@ export default function rotateLogFile(): void {
       fs.statSync(filePath).mtime.getDate() !== today.getDate() &&
       !fs.existsSync(filePath.replace('.log', `_${yesterdayFileFormat}.log`))
     ) {
-      log.info('This log file needs rotation and will be archived');
+      const archiveFilePath = filePath.replace(
+        '.log',
+        `_${yesterdayFileFormat}.log`
+      );
 
       logContent = fs.readFileSync(filePath, 'utf-8');
 
-      fs.writeFileSync(
-        filePath.replace('.log', `_${yesterdayFileFormat}.log`),
-        logContent
-      );
+      fs.writeFileSync(archiveFilePath, logContent);
 
       fs.writeFileSync(filePath, '');
 
-      log.info('Previous log file has been archived due to file rotation');
+      await compressing.zip.compressFile(
+        archiveFilePath,
+        archiveFilePath.replace('.log', '.zip')
+      );
+
+      fs.rmSync(archiveFilePath);
+
+      log.info(
+        `Previous log file has been archived as ${archiveFilePath} due to file rotation`
+      );
     }
   } catch (e: any) {
     log.error('Could not rotate log file: ');
