@@ -1,9 +1,20 @@
+/* eslint-disable react-hooks/exhaustive-deps */
 import { useEffect, useState } from 'react';
+import { ipcRenderer } from 'electron';
 import { Link } from 'react-router-dom';
-import { FiAirplay, FiBell, FiMoon, FiSettings, FiSun } from 'react-icons/fi';
+import {
+  FiAirplay,
+  FiBell,
+  FiDownload,
+  FiDownloadCloud,
+  FiMoon,
+  FiSettings,
+  FiSun,
+} from 'react-icons/fi';
 import '../../assets/styles/components/Navbar/RightButtons.scss';
 import { useTranslation } from 'react-i18next';
 import { useTheme } from '../../contexts/ThemeContext';
+import SpinnerLoader from '../Loaders/Spinner';
 
 export default function NavActionButtons() {
   const { t } = useTranslation();
@@ -11,6 +22,7 @@ export default function NavActionButtons() {
   const [themeIcon, setThemeIcon] = useState(
     theme === 'DARK' ? <FiMoon /> : <FiSun />
   );
+  const [updateActionButton, setUpdateActionButton] = useState(<></>);
 
   function toggleAppTheme() {
     if (document.body.classList.contains('dark-theme')) {
@@ -26,8 +38,52 @@ export default function NavActionButtons() {
     setThemeIcon(theme === 'DARK' ? <FiMoon /> : <FiSun />);
   }, [theme]);
 
+  ipcRenderer.on(
+    'appUpdateStatusChange',
+    (_event: Electron.IpcRendererEvent, { status }: { status: string }) => {
+      if (status === 'AVAILABLE') {
+        setUpdateActionButton(
+          <button
+            type="button"
+            className="optionButton"
+            title={t('navbar.actionButtons.updateAvailable')}
+            onClick={() => {
+              ipcRenderer.invoke('callAppUpdater', {
+                forceOptions: { forceDownload: true },
+              });
+              setUpdateActionButton(<SpinnerLoader />);
+            }}
+          >
+            <FiDownloadCloud />
+          </button>
+        );
+
+        return;
+      }
+
+      if (status === 'DOWNLOADED') {
+        setUpdateActionButton(
+          <button
+            type="button"
+            className="optionButton"
+            title={t('navbar.actionButtons.updateDownloaded')}
+            onClick={() => {
+              ipcRenderer.invoke('callAppUpdater', {
+                forceOptions: { forceInstall: true },
+              });
+              setUpdateActionButton(<SpinnerLoader />);
+            }}
+          >
+            <FiDownload />
+          </button>
+        );
+      }
+    }
+  );
+
   return (
     <section id="rightButtons">
+      {updateActionButton}
       <button
         type="button"
         className="optionButton"
@@ -51,8 +107,8 @@ export default function NavActionButtons() {
       <button
         type="button"
         className="optionButton"
-        onClick={toggleAppTheme}
         title={t('navbar.actionButtons.theme')}
+        onClick={toggleAppTheme}
       >
         {themeIcon}
       </button>
